Fix building for Cygwin with GCC 3.4.4 (Bug 1797). Fix breakpoints with instruction cache enabled (Bug 195).

+++ argtable2/argtable2.c
@@ -969,7 +969,8 @@ void arg_print_formatted( FILE *fp, const unsigned lmargin, const unsigned rmarg
         /* Eat leading whitespaces. This is essential because while
            wrapping lines, there will often be a whitespace at beginning
            of line  */
-        while ( isspace(*(text+line_start)) ) 
+	/* JPB patch to cast argument */
+	  while (isspace ((int) (*(text+line_start))) )
             { line_start++; }
 
         if ((line_end - line_start) > colwidth ) 
@@ -978,7 +979,7 @@ void arg_print_formatted( FILE *fp, const unsigned lmargin, const unsigned rmarg
         /* Find last whitespace, that fits into line */
         while ( ( line_end > line_start ) 
                 && ( line_end - line_start > colwidth )
-                && !isspace(*(text+line_end))) 
+                && !isspace ((int) (*(text+line_end))))
             { line_end--; }
 
         /* Do not print trailing whitespace. If this text
